refactor: moves models file and makes output required
