+++ src/aind_metadata_extractor/models/bergamo.py
